Refactor CMakeLists into subdirectories and reduce redundancy

+++ example/cpp-exchange/dragon/producer.cpp
@@ -17,12 +17,12 @@ int main() {
     timespec timeout{5, 0};
     radex::drg::ddict::Client client{serialized_dd, &timeout};
 
-    client.put_scalar<long>("some-int", 123);
+    client.put_scalar<int32_t>("some-int", 123);
     client.put_scalar<double>("some-float", 1.23);
 
     client.put_tensor<double>("some-float-tensor", {4},
                                       {0.12, 3.45, 6.78, 9.123});
-    client.put_tensor<long>("some-int-tensor", {2, 4},
+    client.put_tensor<int32_t>("some-int-tensor", {2, 4},
                                     {1, 2, 3, 4, 5, 6, 7, 8});
 
     std::cout << "---------------------------\n"
